requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
requests~=2.32.3